refactor(admin): 토큰관련 유틸함수 통합 - get-tokens 함수 삭제 - api.ts에서 토큰관련 함수 통합
