--- apps/admin/src/utils/get-tokens.ts
@@ -1,14 +1,0 @@
-import { getAccessToken, getRefreshToken, removeTokens } from '~/utils/api';
-
-const getToken = () => {
-  const accessToken = getAccessToken();
-  const refreshToken = getRefreshToken();
-
-  if (!accessToken || !refreshToken) {
-    removeTokens();
-  }
-
-  return [accessToken, refreshToken] as const;
-};
-
-export { getToken };
